requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
aio-pika==9.5.4
redis==5.2.1
openai==1.59.3
sse-starlette==2.2.1
python-dotenv==1.0.1
